[trace-vthread] Use static engine headers for trace-vthread-with-static-engine
Adding dependencies on headers causes the underlying library to be added
as a dependency. Thus to add a dependency on trace-engine headers when
using the static engine, trace-engine-static.headers needs to be used.

Tested: readelf of library using static trace-engine previously had a
DT_NEEDED of libtrace-engine.so. Now it doesn't.
readelf -d out/arm64/arm64-shared/libmsd_arm.so

PT-146
PT-195

Change-Id: Ib433303be20a02355994b0d4e4519745280bd7d4