[mt8167][usb] Dynamic endpoint number support.
Currently, the number of implemented endpoint controllers is hard-coded
into the host driver.  This works for the mt8167 chip, but might be
incorrect in a different chip that uses the musbmhdrc IP.  In an effort
to genericize the driver to any chip, we'll read the count of
implemented endpoint controllers from the EPINFO register.

Test: UMS iochk -c 200 /dev/class/block/xxx
Test: runtests -t mt-hci-test-test
Change-Id: I11ddab7da699ce408b840774e209fb2626942120