[fidl][cpp] Bits are represented with class
Far from a fashion statement, representing bits with class rather than
an enum makes it possible to have a conversion operator to a numerical
type (and hence to bool).

This makes it possible to write `if (my_bits_value) { ...`.

A downside is that doing an exhaustive match against a bits value in a
switch statement is no longer helped by compiler warnings... However,
using a bits value in a switch statement and matching against possible
bits members is likely an incorrect pattern in the first place. (This
revealed a bug fostr's handling of bits values, which will be fixed in a
subsequent CL.)

FIDL-620 #done

FIDL-683 #comment no testing infra, tracked in FIDL-623
FIDL-623 #comment test FIDL-683

Test: added some for new behavior, coverage existed for the rest
Change-Id: I922f3e9943d75d44a57dcfc68ee326378fab0ede