[component_manager] Mount with OPEN_RIGHT_WRITABLE
Services need to be hosted with both read and write rights.  There is an
upcoming change for pseudo-fs that will enforce hierarchical rights and
mounting with only read permission breaks, as fdio is already requesting
both read and write when opening a service.

Yefei is working on a smilar change for the C++ code as part of

    https://fuchsia.atlassian.net/browse/ZX-3251

Change-Id: I2bf93638f8ee335b375f9c8deb942d8ec23ffef7