[zircon][libc][fdio] Pass vmar allow faults flag
This change prepares for enforcing the allow faults flag by adding the
flag to the calls in sdk prebuilts which need it.

This relands the userspace portion of ee5c557. That change failed
because some downstream tests currently run newer packages on an older
system, and the older kernel rejects the args passed in by the newer
libc.

Change-Id: I814f51282b0615743605c9871424c35c03705c69