[tools][devshell][aemu] Add software GPU flag
DX-1617 #comment

This provides a flag that disables host side GPU acceleration. This
is useful when running in headless mode.

Vulkan is also disabled when using software GPU as swiftshader
is not yet providing the external memory extensions that are needed
for useful Vulkan support on Fuchsia. This this will change in the
future.

Test: fx aemu --software-gpu
Change-Id: I24f390dd8b9ed394789fc5aeb1efc85eebf08a1b