[netdump] Rename Headers class used by filter to Packet
The filter component expects the caller to pass in an object
representing a packet, which is composed of pointers to headers
as well as the frame length. `Packet` is a more suitable name for
this object than `Headers`. This CL is the find-and-replace.

Bug: CONN-99
Test: Build with `--with-base //garnet/packages/tests:zircon` as a
parameter to `fx set`, then `runtests -t netdump-test` on the target.
This CL does not change any observable behavior.

Change-Id: I812a404d44874a3dfb6007f59158b238eccfcc0c