[inspect] Allow reading from files
This change allows us to pass to `iquery` a path that points to a regular
file.  This is in addition to the current functionality that allowed
us to pass `iquery` a path that points to a VMO or a FIDL interface
of a particuar type.

A limitation is that the path must still end with `.inspect`, as this
represents a hint to the reader that the contents are inspect-formatted.
The limitation exists so that the iquery reader could determine the
correct way to read the data source without consulting the filesystem to
figure out its type. This limitation may be lifted later.

Sample session:

```
$ pwd
/tmp
$ ls -al
dr--------    2 0        0                 0 Jun 17 20:53 .
-rw-r--r--    1 0        0              4096 Jun 17 22:04 file.inspect
$ iquery file.inspect
root:
  version = 1.0
$
```

TESTED=fx run-test inspect_query_tests  # after enabling the respective test disabled due to FLK-186

CF-829 #done

Change-Id: I541276cc729c1cee1243dac9b1ee5e5db5171841