[system-ota-tests] Use the shell to ls/cat files
This breaks our dependency on using the `ls` and `cat` programs,
since it's not guaranteed if these commands will be available
on the system. Instead, this switches over to using shell features
to read files and list files in a directory.

In addition, this change modifies the repository registration to happen
right before starting the OTA process, to avoid potentially downloading
ABI incompatible packages while we validate the state of the device.

Bug: PKG-810 #done
Bug: PKG-811 #done

Change-Id: I934bfddf51f3d4751554578731676625fe55a79d