[pseudo-fs] Overhaul of access checks.
Turns out I have missed the access checks for the `Open` FIDL call. I
wanted the `open()` call on the files and directories to be unrestricted
by access checks, as this method is the internal server API.  For
warding `Open` FIDL call directly to the `open()` method was skipping
the access checks.

As some of the tests were broken and as these are negative checks, this
issue was not observed by any of the working unit tests.

Plus changes in the access rules (directories now having proper
OPEN_RIGHT_WRITABLE support) requred a few additional changes to the
unit tests.  In particular,

    directories_do_not_restrict_write_permission

is no longer a meaningful test, as directories now do restrict write
permission.  Changed it to check that persmissions are indeed
restricted.

On top of the above, it appears a number of other checks were also done
incorrectly, in particular, when we have added hierarchical access
permissions, we did not run all the tests.  And OPEN_FLAG_POSIX was not
checked correctly for directories either.

Change-Id: I142968509e35830491ffb91c64d6dd4067b6c6a0