[tools] fx-command-run works for contrib scripts
As discussed, fx-command-run should support the same scripts found by
"fx", which includes scripts in "contrib". This change adds a bash
function "fx-find-command" (similar to the "find_command" function in
the "fx" script, but only checks devshell and devshell/contrib), and
uses this function to find the script for fx-command-run.

Test: tested by hand
Change-Id: I15f1472dd4753503c60cae4d61b7bb7375d248be