[dash] USE_GLIBC_STDIO, fixes many flushing bugs
At one point in time, this code path probably was not viable, but it works
just fine today.

There were two options to fix the builtin functions in zircon.c, either to
teach zircon.c to use output.h only (instead of stdio.h), or, teach dash to
use stdio.h. I opted for the latter, as the behavior of Dash's non-stdio
buffering is broken in various ways I could not obviously fix.

Test: /system/test/sys/dash-test
Change-Id: I8e460520464d2340502ba0433bda909d24b8ba50