[fidl][llcpp] Nest all generated types under ::llcpp
For example, given the following FIDL definition:

    library foo.bar;
    struct S { int32 a; };

The generated struct will be placed under the
llcpp::foo::bar namespace.

This is necessary to avoid symbol collision with the high-level
C++ bindings, whose coexistence with LLCPP is expected as LLCPP starts
being used in filesystem servers.

In addition, if a FIDL library name starts with the component "llcpp",
the LLCPP codegen would rename it to "llcpp_", to avoid confusion with
the outermost llcpp namespace.

TEST: builds and added new cases in fidlgen_cpp_ir_test

Change-Id: Ie6f9933fc2df7f9b66644a4f385e84ce505a1725