[netdump] Clean up argument passing for various filters
Clean up argument passing on construction of port, and MAC and IPv6
address filters.

For the port filter, the input port range vector is passed by value,
the data is massaged in place, and the vector is moved into the
filter. This minimizes the amount of copying required.

The storage types of MAC and IPv6 address are changed to `std::array`
instead of the current C-style arrays, for the following reasons:
- C-style arrays are discouraged in C++ in favor of `std::array`.
- Readable typedefs can be used, such as `MacAddress`, instead of a
  raw array type.
- Code becomes simpler and more readable, e.g. copy assignment and
  copy constructor can be used instead of `std::copy`.
The address data in packet header structs must remain as C-arrays
for compatibility.

Bug: CONN-99
Test: Build with `--with-base //garnet/packages/tests:zircon` as a
parameter to `fx set`, then `runtests -t netdump-test` on the target.

Change-Id: I05d165f11291bc9b95358c641f168991235a7c75