[gtest,audio] avoid polling for conditions
This CL changes RealLoopFixture.RunLoop[WithTimeoutOr]Until to accept
an infinite step value, in which case the loop blocks normally,
checking the condition when in unblocks naturally. This is useful in
90% of cases - cases in which the condition becomes true as the
result of executing a task on the loop. This eliminates the need for
polling the condition at some fixed 'step' frequency (the default
being 10ms).

Audio tests are configured to use an infinite step value.

TEST: audio tests
Change-Id: I1d4640f4101a28d761603ba77b772f85ebd0e72e