[netdump] Netdump filter parser external interface
This CL is part of a series that adds packet filtering capability
to netdump using a filter language with similar syntax to the Berkeley
Packet Filter (BPF).

This CL defines the public interface of the filter string parser.
There is an interface to client code that wishes to use the parser,
as well as a builder interface that will interact with specific
implementations of packet filter objects. Internal structures in the
parser that track state during a parse are also defined.

Follow-up CL will include:
- Syntax logic for the filter language.
- Adapter of the parser interface to the checked-in filter tree
  nodes implementation.

Bug: CONN-99
Test: Build with `--with-base //garnet/packages/tests:zircon` as a
parameter to `fx set`, then `runtests -t netdump-test` on the target.

Change-Id: I2475206768c11f809be0e8b949d299aa773cfa92