[testsharder] Extend testsharder to take a test multipliers json file.
This file should contain tests that are to be run multiple times which the testsharder
will use to create a separate shard per test that contains the same test written as many
times as the total_runs specified in the file.

The output will contain the original shards it currently produces plus
these new shards with the multiplied tests. This will be used to test for flakiness.

Bug: IN-1231 #comment
Change-Id: I4366ff012bc929c938c6aeaa9a26812e7888c7ef