[kernel][vm] Make closing COW clones more efficient
In general, it is expected that bidirectional COW clones share most of
their pages with their parent. Given that, it is preferable for closing
a clone to do work proportional to the number of pages forked, rather
than something like the total number of pages committed in the parent.

To accomplish this, this change first replaces the unconditional
release of ancestor pages in the dtor with more selective work when
cleaning up the hidden parent, as the merge operation had been undoing a
lot of the release work. Second, this change adds a fast path for
merging the hidden vmo and its child that applies to some of the common
use cases for COW clones.

ZX-1268 #comment [kernel][vm] Make closing COW clones more efficient

Change-Id: I49ad6c9fed64fc6edc03b812b44189b3087dab7f