[audio_core] Enforce payload_buffer_id.
At the moment, audio_core only supports a single audio payload buffer.
The SubmitPacket API was not checking that the client properly
specified the payload buffer was properly provided in the StreamPacket.
This simply adds this validation and will disconnect any clients that
provide an invalid buffer ID.

This change is to prepare for a following change where we will enable
clients to register more than one payload buffer with an AudioRenderer.

Test: fx run-test audio_fidl_tests
Change-Id: Ia7dd32c3d2fbe33cb8154ef989f870b42f629bbb