[netclock] Initial implementation of fuchsia.time.Utc.
Adds two new FIDL protocols and the netclock service which serves
one and consumes the other. Includes a single basic test to
demonstrate functioning update notifications.

* fuchsia.time.Utc

Notifies clients that changes to UTC have occurred, and in the
future will vend access to system time.

* fuchsia.time.DeprecatedNetworkSync

Until netclock can assume `network_time`'s responsibilities to...
actually procure time from accurate sources, we must be notified
when this other process changes the value of the UTC offset. This
API is insta-deprecated, as it only exists to support a temporary
migration of features from one service to another. We will delete
it in a future change.

Change-Id: Icbc9c5c0c7d1022a57b6f6a75762e8c4435ba5bb