[scenic] Apply clipping bounds with view properties
Previously ViewProperties were only being used for clipping on
ViewHolders if the ViewHolder already had a view_ set. That causes
clipping to be incorrectly missed in some cases, and the
SendViewPropertiesChanged() event is always fired after view addition in
LinkResolved for cases like this. Also adds a unit test to cover this
edge case.

Change-Id: Idbfafca271bacdbf5e184a7609074529d74b7574