[debugger] Improve GCC symbols.
GCC encodes static arrays differently in DWARF. Allow both "upper bound"
and "count" for the size. This fixes a decode error for GCC symbols.

Adds a test for the unusual DWARF encoding of __func__ that GCC
generates. The code handled this correctly but seems worth testing since
I found this surprising.

Reads the "artificial" tag from DWARF and skips these variables in
printing in some cases. Adds a test for structure printing with these
variables. This ends up hiding the generated __func__ variables that had
the bug above.

Addresses a theoretical problem in array printing where it wasn't
getting the concrete type for the array values before decoding. I think
our compilers will generate real and not forward-declarations for this
so it may not ever be encountered.

Shorten the error messages for optimized out and unavailable error
messages. These appear a lot in locals and backtraces and the long
messages makes everything harder to read.

Change-Id: Idc4f73d7d0c2e0fffaa7c84afa9336a0f3ebba39