[test_random] do not truncate the seed
When initializing std::independent_bits_engine directly with an integer
value, it is treated as the output type of the generator. This caused
the seed to be truncated to 8 bits. Instead, pass the seed directly to
the underlying generator.
This change modifies the sequence of value generated for a given
TEST_LOOP_RANDOM_SEED: it should be reverted when attempting to
reproduce bugs with older seeds.

LE-798 #done

Test: TestRandomTest.NoSeedTruncation fails without this change
Change-Id: Ia426186dcd549346dda3c917c8febaeff2d42c2a