[activity-service] Add fuchsia.ui.activity FIDL interface.
fuchsia.ui.activity is the API of the ActivityService, which is the
source of truth for the activity state of the system. In a nutshell,
this service is responsible for answering the question, "is the system
currently being used?"

Activity is defined based on recent interaction between the system and
its environment. A few possible examples of activity:
  * A user recently pressed a key on the keyboard, or moved the mouse
    cursor.
  * A user is watching a video or listening to an audio file.
  * A user-defined alarm recently went off.

This CL defines two possible activity states: IDLE or ACTIVE. The exact
mechanics of how these state transitions occur is an implementational
detail of the ActivityService, but the general idea is that a system is
ACTIVE if it had recent activity, and IDLE otherwise.

Two FIDL protocols are defined in this CL, which collectively form the
ActivityService.

  fuchisa.ui.activity.Tracker receives evidence of activity from the
  system and drives the ActivityService's state machine.

  fuchsia.ui.activity.Provider is an observable interface which notifies
  clients of changes to the system state.

Change-Id: I9eee5173a886b7f63ea53710cb7f597320ecbdb8