[identity] Google authorization flow using Chromium
This adds an interface for starting a Chromium context and frame,
and listening for events from the frame.  This also adds logic to
direct a Chromium frame to the Google OAuth login and pull an
authorization code out of a redirect event when login is complete.

Testing: Unit tests, manually log in and verify that auth code length
log is present.  Unit tests are not added for complete interations
with Chromium to avoid creating complex fakes.  Some coverage for
these interactions will come from end to end testing.

Change-Id: I361c8a6e9959cc187ba5bdd09ffe9fcbe201abb2