[storage/fs-management] add rust library for filesystem management
This library provides similar functionality to the fs-management library
in zircon, except with a rust api. This support is done in a very
similar way to the fs-management implementation. It has support for
mount, unmount, fsck, and format.

Currently this library only provides support for blobfs partitions.
minfs partition support is planned as well. (ZX-4302)

ZX-4298 #done
ZX-4300 #done

Test: fx run-test fs-management-test
Change-Id: I41793432e24d38b0bea2f53f7981361fd19fe517