[fidl] Support strict mode for xunions in fidlc.
This CL adds the following to fidlc:

* parsing the `strict' keyword for xunions,
* adding a strictness flag to the flat & raw ASTs for xunions,
* emitting a "strictness" member to the JSON xunion object,

This is part one of many CLs to fully support strict mode for xunions,
tables, bits and enums. This CL intentionally supports xunions only;
support for other types and the rest of the FIDL toolchains will land
soon.

There are also minor changes to FIDL's CONTRIBUTING.md to update it for
fx workflows.

Test: fx build zircon/system/utest:host && $FUCHSIA_DIR/out/default.zircon/host-x64-linux-asan/obj/system/utest/fidl-compiler/fidl-compiler-test.debug
Test: fx run-test fidl_tests  # run C++ tests on device
Test: fx run-test go_fidl_tests  # run Go tests on device

FIDL-523 #progress

Change-Id: I0bcd1f076370f7f9c8683c0ca2c6033ebf08e470