[media][tools] signal_generator can manipulate timestamps
The signal_generator utility now has the ability to manage
the presentation timestamp (PTS) applied to each buffer that
it sends to the audio renderer. It also has the ability to
set its continuity threshold, a value used by the renderer
when deciding whether to honor the PTS as-is, or (if the
delta is less than the threshold) to align the buffer
"gaplessly" immediately after the previous one.

Developer tool, not a production service

Change-Id: Iae9639b1fce30155f05a26e25fe9d799a60f6b73