[banjo] add support for fidlcat generation
Generate data to be able to decode all the syscalls.

This data is then used by fidlcat to display all the
arguments of some or all the system calls of a process.

Currently, around 80% of the arguments can be displayed
correctly.
To be able to display 100%, we will need to add some information
within the syscalls.banjo file.

Change-Id: I7657b2cbfc08fa37d5d741e1cecbd233fc8d84e2