[devcoord][factory] Route fuchsia.boot.FactoryItems to svchost.
ZX-3611 #comment
Adds fuchsia.boot.FactoryItems to list of services that are re-routed to
bootsvc.
Updates svc-test to check the service connection.
Service needs to be exposed in appmgr which is handled by fxr/280609.
Updates root manifest to offer FactoryItems to devcoordinator.
Once that lands, this service will need to be added to the list of
additional services in the realm for components v1.

Test: cd zircon && ./scripts/build-zircon-arm64 && ./scripts/run-zircon-arm64;
      In QEMU, /boot/test/ddk/driver-test-test

Change-Id: I777984e15cde8ce9fe71bf74b26f75ad6b391c99