[hermetic-compute] Revamp launching with no engine-side assembly
The controlling process completely controls the register state by
starting the process with its thread suspended and then mutating its
registers before letting it run.  This makes it possible to do all
the register and stack setup precisely from the controlling process
according to the C ABI so there is no startup code in assembly.

This substantially revamps the high-level launching API.  The
"fluent" Launcher object API is replaced with a single call where
special tag types in the argument list direct export agents to
perform all the setup tasks as side effects.  The handle and vDSO
are no longer special cases but instead are normal arguments with
export/import behavior like any other.  The HermeticComputeEngine
class implicitly expects a vDSO argument to be its first argument,
but that is not required by the framework.

Change-Id: I60c0b1f5087f31b1d461ebab18c238d8d21384cd