[ddk][devmgr] Handle unimplemented ops for device API
This change gracefully handles the cases where device_get_size,
device_read, or device_write are called without the corresponding op
being implemented by the driver.

Also add a test for the API. Check that we handle the above case, as
well as correctly routing the device context and arguments.

ZX-4264 #done

Change-Id: I5404ce0e26dd9acb492f2e8df94a063438a535c6