[usb-fwloader] Use cast to avoid error due to ambigous conversion
libc++ has recently implemented P0608R3 (http://wg21.link/p0608) which
results in an error when narrowing conversions and conversions to bool
occur while assigning to variant. This is the case for usb-fwloader
since fuchsia_hardware_usb_fwloader_PrebuiltType_* constants use
UINT8_C which could be converted to const char * or uint8_t which is
the underlying type of fuchsia_hardware_usb_fwloader_PrebuiltType.
We use explicit static cast to avoid that error.

Change-Id: I01e43cdf30c43d5e3d0092f18b6707faf12fb3e2