[fidlc] Correcting non-deterministic topological decl ordering
We used to rely on 'name ordering', which is not deterministic because
it relies on the pointer address of the libraries in which the two
compared names are. Relying on the pointer address is required in the
default ordering since library names are compiled after the fact.

When doing topological sorting, all library names have been compiled,
and we can therefore rely on those to provide a fully deterministic
ordering.

Verified this fix against reported issue in fuchsia-camera-common as
follows (on JIRI_HEAD and with this patch):

(1) Generate an expected.json

    ./build-zircon/tools/fidlc \
        --json expected.json \
        --files system/fidl/fuchsia-sysmem/allocator.fidl \
                system/fidl/fuchsia-sysmem/collection.fidl \
                system/fidl/fuchsia-sysmem/collections_deprecated.fidl \
                system/fidl/fuchsia-sysmem/constraints.fidl \
                system/fidl/fuchsia-sysmem/driver_connector.fidl \
                system/fidl/fuchsia-sysmem/format_modifier.fidl \
                system/fidl/fuchsia-sysmem/formats_deprecated.fidl \
                system/fidl/fuchsia-sysmem/heap.fidl \
                system/fidl/fuchsia-sysmem/image_formats.fidl \
                system/fidl/fuchsia-sysmem/image_formats_deprecated.fidl \
                system/fidl/fuchsia-sysmem/usages.fidl \
        --files system/fidl/fuchsia-camera-common/common.fidl \
                system/fidl/fuchsia-camera-common/virtual.fidl

(2) In a loop, generate an actual.json (similar fidlc invocation), and
diff expected.json actual.json

On JIRI_HEAD, differences occur every so often. With this patch, output
is deterministic, and no differences ever occur.

FIDL-673 #done

Test: see decls_across_libraries, which fails about 4% of the time on
OSX but is always green on Linux

Change-Id: I52be4d1c27887f10c720391290040500940d6ccc