[benchmarks] Update package reference: garnet_benchmarks -> fuchsia_benchmarks
The code in this benchmarks.cc file used to live in garnet_benchmarks,
but it is now in fuchsia_benchmarks, so the latter is the more correct
reference.

This will allow the perf-test executable to be removed from the old
garnet_benchmarks package, once benchmarks_perfcompare.cc is also
moved to fuchsia_benchmarks.

Issue: PT-175
Test: "peridot-x64-perf-dawson_canyon" trybot on the CQ
Change-Id: I8fbf722abc9cf0f919c67bdbfef644a3ecf20489