[astro][brcmfmac] Fix asan issue in brcmfmac driver
The Amlogic SDIO driver reads packet data four bytes at a time.
brcmf_sdiod_netbuf_read accounts for this by rounding up the receive
size, however brcmf_sdiod_ramrw was not allocating enough memory to
to make up the difference. The result was a heap overflow when reading
packets with a size not divisible by four.

Test: No crash seen booting Astro with bringup/asan build.
Test: Can scan, connect to WiFi on Astro and Sherlock with core build.
Change-Id: I31d21475cd8fce35d83197bce3c679d2f7e0b566