[devmgr] De-C++-ify two C++ symbols in driver.so
There were two last functions exported in C++ style:

    devmgr::getenv_bool(char const*, bool)
    devmgr::device_host_main(int, char**)

For getenv_bool, it looks like it is not used outside of driver.so,
hence it is hidden by making the default visibility of the corresponding
source_set hidden.

For device_host_main, it is used externally by devhost, hence a wrapper
devmgr_device_host_main with C linkage was added.

This unblocks landing the symbol validation change:
https://fuchsia-review.googlesource.com/c/fuchsia/+/290743

Change-Id: I29995a4384a57b0cb74b5f47d7766ba77f8b3f35