[netstack3] Add ICMP connections
- ip::icmp:
  - Add IcmpState
  - Add IcmpEventDispatcher trait with receive_icmp_echo_reply
    method
  - Add ICMP connections
  - Fix bug in Echo reply for ICMP v6.

NOTE: this is the first push for having ICMP connections and the
implementation is expected to evolve with the addition of a new sockets
API. For now, ICMP connections are only bound to remote addresses and
the ICMP identifier. Future changes will change that behavior to a more
complete and correct one.

Change-Id: Ibadc4b3452ee7da87a758fd3e793f4cb48faf50d