[scenic] Additional ViewBound PixelTesting
Further unit testing to ensure that when views are placed
underneath another node that has a transform applied to it
already, that the view bounds themselves are transformed
appropriately even if they are specified by the client in
local coordinates. (E.g. If a view bounds minimum position
is specified at (0,0,0) and the view is underneath a node
that is translated by (5,5,5) the view bound's minimum
coordinate should be at (5,5,5) in world-space.

Change-Id: Ib1fd7ce3add49b04c3342246a4bf228339dbc28c