[scenic] correctly schedule frames by presentation_time
Scenic clients have the option to specify a presentation_time when they
call Present(). Because Scenic was not inclusive of the upper bound of a
target_presentation_time however, when presentation_time ==
target_presentation_time we would end up scheduling the Client's frame
on the next Vsync instead of the current, leading to a dropped frame
every Present().

SCN-1479 #done

Test: unit tests, running apps that specify a non-zero presentation
  time, such as BaseView - spinning square running at 60fps

Change-Id: Id96ae9923b464113ba336bc0a185a2f931f17ccf