[debugger] Add an async output buffer.
This is a general version of the asynchronous output handling in
FormatValue. It will allow us to generate different types of
asynchronous output, not just for value formatting.

The new value formatting system will use this scheme and the old one
will be deleted.

Hooks it up to one place: the memory formatter. It doesn't particularly
improve readability in this case but it has the advantage of having unit
test coverage for end-to-end async console output.

Change-Id: I65d0870f5c16d898e499556bfbad9430cc87caea