[mediaplayer] change renderer end-of-stream logic
This CL changes the way the audio and video renderer nodes deal with
end-of-stream in order to address test flakes.

For both audio and video, end-of-stream is indicated in the stream by
a packet with an end-of-stream flag. That packet has a PTS, which can
be correlated to a real time (in CLOCK_MONOTONIC) using the timeline
function. This calculation was used previously, in combination with
a delayed task, to determine when a renderer node should signal end-
of-stream. This approach assumes that the actual renderer (ImagePipe
or AudioRenderer) renders frames in a timely manner.

This works fine in normal production situations. In normal operation,
the ImagePipe or AudioRenderer renders in a timeline manner, and any
discrepancy is typically not apparent to the user. In CQ, however,
the constrained environment encountered in some test runs means the
ImagePipe or AudioRenderer, though fake, don't function promptly. In
such cases, end-of-stream may be signalled e.g. before the fake
AudioRenderer has received and processed the last payload-bearing
packet. Some CQ tests verify that all expected payloads have been
rendered when end-of-stream is signalled, so flakes may occur.

This CL changes the logic concerning the timing of the end-of-stream
indication. Rather than calculating a real time and using a delayed
task, the renderer nodes instead wait for all expected packet
releases to occur, tracking the progress of releases and signalling
end-of-stream when all packets have verifiably been rendered. In
production situations, this may delay end-of-stream signalling by
a few milliseconds, but the effect is negligible. In CQ tests, this
approach ensures that e.g. the fake audio renderer will have
received and processed all payloads before end-of-stream is signalled
and the fake renderer is interrogated to make sure expectations have
been met.

TEST: MediaPlayerTests::PlayWavDelayEos (in CQ) explicitly tests that
      end-of-stream is signalled on audio renderer packet retirement
      rather than on intended end-of-stream time. The test fails
      without the renderer changes and succeeds with them.

FLK-161 #done

Change-Id: I888dc70b573b361b0b80932830b965050fbeab17