[identity] AccountHandler supports ephemeral
Here we add support for ephemeral accounts in AccountHandler, by
creating a separate component `account_handler_ephemeral`, which does
not have persistent storage privileges. The flow
of the ephemeral bit starts at the execution of the binary with an
--ephemeral flag, and then
lives on the instantiations of AccountHandler, Account, Persona and
TokenManager.

The AccountHandler, Account and Persona gets a new enum AccountLifecycle
which contains the path to the account dir if the account is persistent.
TokenManager already supports in-memory auth db, so here we just use
different constructors depending on the context.

Future work:
- Support in AccountManager.

AUTH-76 #comment

Change-Id: I6761dd6d5888db7ce2ec318acb2615f1ef1286e1