[fuchsia-watch] Recursive directory watching in Rust
This change implements a recursive directory watcher in Rust for
Fuchsia. This is higher-level than fuchsia-vfs-watcher. While
that crate provides a low-level wrapper around fuchsia.io FIDL calls
for watching, this crate provides enums to describe operations in terms
of PathEvents that contain typed Nodes.

The interface simply takes an individual absolute path to watch, and it
returns events containing the whole absolute path at which an event
occurred.

Change-Id: I0dc108c363029ddc43cd0446a13ef4e57826ed0f