[vim][eth] Disable perf counters
The dwmac in the S922 contains ~75 different performance counters
for monitoring state of the MAC.  By default, these counters generate
interrupts when they are at 1/2 max and max value. The octet counter
was the first to overflow the 2^31 halfway point, resulting in a
interrupt flag being set and not cleared.  This caused the top level
interrupt for the mac to not be re-armed and as a result the MAC stops
handling packets.

This patch disables all the counter interrupt sources.
JIRA: ZX-4266
Test: manual test 'sudo ping6 -f -l 50 `out/default.zircon/tools/netaddr`
and ran for 12 hours to confirm none of the overflowing counters would
stall the mac.

Exhaustive tests for this would require generating conditions to
overflow all the counters, most of which are 32-bit and capture very
specific line conditions which is impractical for automated
testing.

Change-Id: I71ef374cf7196af6abdc256527df4d86b930b292