[magma] Implement delayed sizing of buffers
The sysmem API is designed to allow multiple different sizes of images
to be allocated in a buffer. This can prevent needing to allocate
new memory on video stream resize.

To use this feature, we need a way to specify the actual size of the
image that is used (this will be given through vkCreateImage), which
overrides the normal size used to create the image.

MA-638 #comment

Change-Id: I44ba430ca330df478e48f08c133ec1635f3e9d02