[zircon/blobfs] Refactor the JournalProcessor state machine.
The state machine is now explicit and much simpler. It's part of
JournalEntry, which still has a way to go in terms of simplification
(work in progress).

This CL also removes the last queue from the journal (sync_queue_),
so sync entries are sent to the write back queue when normal entries
have been deleted, but before the journal info header is updated.

ZX-2925 #done

Change-Id: I51e1846682465d601f704d060290d854221cfcfb