[sdk] Validate golden symbols for DSO prebuilts in the SDK
Since the CL to add vulkan.symbols.abi to third_party/vulkan landed,
we can also now start validating that the symbols do not change.

Static libraries are more complex and likely requires a more complex
policy on public symbols, and after syncing with the Toolchains team,
we decided not to address them in this patch.

DX-688 #comment

Change-Id: I71e2bc0024fa6c5cda2bdec5077c42322357f9b4