[tools][devshell][aemu] Add HTTP flag
This adds a simple way to enable remote access. gRPC port will
default to 5556 and all that needs to be specified is the HTTP
port used for gRPC control messages and WebRTC traffic.

This also launches the emulator as a child process so 'trap'
works correctly. Temporary files and child processes are now
cleaned up properly.

Test: fx aemu -x 8080
Change-Id: I74ebdf6f96a3c61365f8633594706a7f7045784f