[vmm][tests] Run loop before starting devices
The service directory runs on the same async loop as the test, so it's
possible for the device sync pointers to begin waiting on its channel
before the other end has been connected to the device process and cause
a deadlock. We can ensure that the connection happens first by running
the loop until idle after the call to Connect.

ZX-4395 #done

Change-Id: I2c3d84c9645e43e2327899a5af80b7449d343749