[virtualization] Add logging for long-running operations.
Some operations, such as connecting to guest serial or waiting for guest
VMs to start have long timeouts. When an operation fails, however, we
proceed to just keep polling until the timeout, with very little
feedback is given to the user, making it hard to diagnose.

Add some logging that will print a message every k seconds for
long-running operations to assist with debugging.

Change-Id: I9b86c1cd1301126be4087c3347beeb740320077c